+++ HISTORY.rst
@@ -29,6 +29,7 @@ End-User Summary
   SODAR can also be queried for phenotypes.
 - Properly sanitized output by Exomiser.
 - Rebuild of variant summary database table happens every Sunday at 2:22am.
+- Added celery queues ``maintenance`` and ``export``.
 
 Full Change List
 ================
@@ -54,6 +55,7 @@ Full Change List
 - Adding files for building Docker images and documenting Docker (Compose) deployment.
 - Properly sanitized output by Exomiser.
 - Rebuild of variant summary database table happens every Sunday at 2:22am.
+- Added celery queues ``maintenance`` and ``export``.
 
 -------
 v0.22.1